Move OAuth definitions into shared, add more
